+++ okta-identity-dart-auth-sdk/lib/okta_identity_dart_auth_sdk.dart
@@ -1,0 +1,33 @@
+// lib/okta_sdk.dart
+
+// Base SDK Setup
+export 'src/base/okta_issue_okta_base_sdk_setup.dart';
+
+// Authentication Flows
+export 'src/auth/okta_issue_okta_auth_login.dart';
+export 'src/auth/okta_issue_okta_social_login.dart';
+export 'src/auth/okta_issue_okta_oidc_logout.dart';
+export 'src/auth/okta_issue_okta_saml_logout.dart';
+export 'src/auth/okta_issue_okta_token_exchange.dart';
+export 'src/auth/okta_issue_okta_token_validation.dart';
+export 'src/auth/okta_issue_okta_global_token_revocation.dart';
+
+// User and MFA Management
+export 'src/user/okta_issue_okta_user_management.dart';
+export 'src/user/okta_issue_okta_multi_factor_verify.dart';
+export 'src/user/okta_issue_okta_authenticator_management.dart';
+
+// Metadata
+export 'src/metadata/okta_issue_okta_metadata.dart';
+
+// Single Sign-On
+export 'src/sso/okta_issue_okta_idp_initiated_sso.dart';
+
+// Client Registration
+export 'src/registration/okta_issue_okta_dynamic_client_registration.dart';
+
+// Authorization
+export 'src/authorization/okta_issue_okta_authorization.dart';
+
+// Utilities
+export 'src/utils/okta_issue_okta_utility_methods.dart';